added styling with Material UI

+++ src/components/GalleryList/GalleryItem.jsx
@@ -1,6 +1,10 @@
 import { useState } from "react"
 import axios from "axios";
 import "./GalleryList.css"
+import Swal from "sweetalert2";
+import { IconButton } from "@mui/material";
+import FavoriteIcon from '@mui/icons-material/Favorite';
+import DeleteIcon from '@mui/icons-material/Delete';
 
 export default function GalleryItem({ getGallery, item }) {
 
@@ -11,9 +15,26 @@ export default function GalleryItem({ getGallery, item }) {
     }
 
     const deletePost = (imgId) => {
-        axios.delete(`/gallery/${imgId}`)
-            .then(() => getGallery())
-            .catch((err) => alert(err));
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "This post will be deleted forever!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Deleted!',
+                    'This post has been deleted.',
+                    'success'
+                )
+                axios.delete(`/gallery/${imgId}`)
+                    .then(() => getGallery())
+                    .catch((err) => alert(err));
+            }
+        })
     }
 
     const [showDescription, setShowDescription] = useState(false);
@@ -32,15 +53,19 @@ export default function GalleryItem({ getGallery, item }) {
 
 
     return (
-        <div className="galleryItem" key={item.id}>
-            {showDescription
-                ? <><div style={backgroundStyle} />
-                    <p className="caption" onClick={() => setShowDescription(false)}>{item.description}</p></>
-                : <><img  src={item.path} onClick={() => setShowDescription(true)} /></>}
-            <p>{item.likes} likes</p>
-            <button onClick={() => likeCounter(item.id)}>&hearts;</button>
-            <button onClick={() => deletePost(item.id)}>X</button>
-        </div>
-    )
+            <div className="galleryItem" key={item.id}>
+                {showDescription
+                    ? <><div style={backgroundStyle} />
+                        <p className="caption" onClick={() => setShowDescription(false)}>{item.description}</p></>
+                    : <><img src={item.path} onClick={() => setShowDescription(true)} /></>}
+                <p>{item.likes} likes</p>
+                <IconButton size="small" color="error" onClick={() => likeCounter(item.id)}>
+                    <FavoriteIcon fontSize="small" />
+                </IconButton>
+                <IconButton aria-label="delete" size="small" color="secondary" onClick={() => deletePost(item.id)}>
+                    <DeleteIcon fontSize="small" />
+                 </IconButton>
+            </div>
+            )
 
 }
